GUI windows added
When the starting prompt menu is opened clicking ehither button opens the next window in the process. However with the adition of the mainloop in each of the adddtional classes the destory of the starting menu stopped working. Fix is not needed, least for now, main menu can just remain open.